pyproject.toml
[tool.poetry]
name = "data-quality-pydantic"
version = "0.1.0"
description = ""
authors = ["Aurelio <aurelio.boelter@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
mkdocs = "^1.6.1"
apache-airflow = "^2.9.0"
pytest = "^9.0.1"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_third_party = ["airflow", "pytest"]
skip_glob = ["*/__init__.py"]

[tool.ruff]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501", "D", "ANN101", "ANN102"]
line-length = 100
target-version = "py311"
fix = true

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "D103"]  # ignora regras mais rígidas em testes
"dags/*" = ["B905"]           # ignora warnings específicos em DAGs longas

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.4.0"
ruff = "^0.14.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
